+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Added
 - Test cases for pyrvapi metadata object added
 - Test case for ``latticesearch.pdb_in_results`` added
 - ``-tab_prefix`` option added for JScoFe
+- ``CCP4`` container for related information
+
 Changed
 ~~~~~~~
 - Removed reference to deprecated module ``iotbx.pdb.mining``
@@ -22,6 +24,8 @@ Changed
 - Bug fix in ``simbad.lattice.latticesearch`` for duplicate entries from alternative unit cells
 - Bug fix for logging and error message handling prior to logger initialisation
 - Bug fix in ``simbad.util.pdb_util`` variable name 
+- ``ccp4_root`` function changed to ``CCP4RootDirectory`` class
+- ``ccp4_version`` function changed to ``CCP4Version`` class and version extracted from official CCP4 release file
 
 0.1.0
 -----